fix(#2827): clippy clean + split partition_access types into their own file
Splits the closed label sets / AccessWeight / WindowSummary vocabulary out of
partition_access/mod.rs (834 lines) into types.rs so both stay inside the
campsite ratchet, and clears three lints: a single-arm match, a useless
into_iter, dead KeySizeNote variants under the tombstones build, and a std
MutexGuard held across an await in the e2e test (now tokio::sync::Mutex).
